+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ### CHANGELOG
+#### chameleon-bridge@0.2.2
+- 支持tt和quickapp
+
 #### chameleon-bridge@0.2.1
 - 修复callNative交互，weex IOS端传递参数给js中的数据对象包含？&,导致split错误